Makes this package compilable, but does not link due to missing symbol.
git-svn-id: https://svn.aros.org/svn/aros/trunk@17648 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/aminet/comm/term/xprz351/Utils.c
@@ -34,11 +34,11 @@
   * Version 3.52  June 1995 , fix incorrect values for certain setup options
   **********************************************************************/
 
-#include "xprzmodem_all.h"
 
 #define CATCOMP_NUMBERS
 #include "xprzmodem_catalog.h"
 
+#include "xprzmodem_all.h"
 
 /*
  * Transfer options to use if XProtocolSetup not called
@@ -640,7 +640,7 @@ XProtocolHostMon (
       ++sv->matchptr;        /* string, increment match position */
       if (!*sv->matchptr)
       {                        /* if at end of match string, it all matched */
-        sv->buflen = actual - (sv->bufpos - serbuff);
+        sv->buflen = actual - ((long)sv->bufpos - (long)serbuff);
         XProtocolReceive (xio);
         sv->matchptr = startrcv;
         actual = 0;
